Allow artefact suffix names to be passed in as argument
It is possible that an artefact repository has a different name from the artefact itself due to artefact naming conventions. For example scala-js web jar artefacts naming convention is to add a suffix of '_sjs' with the version of the scala-js plugin, i.e. '_sjs0.6'. This change will optionally allow for a suffix to be passed in